# Optimize conversion of Key enum from/to string, reduce allocations